+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+- Added support for publishing packages from the default branch.
+- Added support for verifying packages from PR branches.
+
 ## 0.1.0
 
 - Initial version.